✅ test: extract MockFramer to internal/testutil and update tests
- Add internal/testutil.MockFramer (SentFrames, SendError, and helper frame creators)
- Replace duplicated mock framer implementations in basic_test.go and vhost/recover_test.go with testutil.MockFramer
- Update tests to use exported SentFrames field and to inject SendError for failure scenarios
- Remove now-unused local mock types and unused imports
